Move lora to radio library
