+++ lib/f_core/radio/CMakeLists.txt
@@ -1,0 +1,7 @@
+# Copyright (c) 2025 RIT Launch Initiative
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_library()
+FILE(GLOB sources *.cpp)
+zephyr_library_sources(${sources})
+
